Add Dockerfile to facilitate reproducible builds.
We want people to be able to reproducibly build their own ia2 images.
This Dockerfile makes that process (reasonably) easy.  We base the image
off of Amazon's Linux, install the Nitro command line tools, and then
compile the ko image that's in our local docker registry.  To be able to
access our local docker registry, we need to pass the socket to our
local docker daemon to the image.