Seed entropy pool before initializing modules.
The chi module appears to need cryptographically-strong random bytes
when it initializes but that blocks because we haven't seeded our system
entropy pool by then.  This patch moves getNSMRandomness to a separate
module whose init function is called before chi, to ensure that the
entropy pool is initialized by the time chi needs random bytes.