Add GitHub action and reproducible Dockerfile.
This commit adds ia2 to our general docker build pipeline by adding a
GitHub action.  The commit also adds a Dockerfile that allows for
reproducible, kaniko-based builds.  Simply run "make docker" to create a
reproducible image.

We currently use kaniko in version 1.7.0 because 1.8.0 (the newest
version as of 2022-03-22) has a bug that breaks reproducible builds:
https://github.com/GoogleContainerTools/kaniko/issues/2005